Move integration tests under integration build tag (#5)
* Move integration tests under integration build tag

Moves integration tests into separate files so that they don't get run in CI. This speeds up the `make test` command significantly
Integration tests can be run with the `make integration` target

* Drop required coverage to 82%

* Fix unused var lint issue
