+++ Makefile
@@ -9,11 +9,14 @@ clean::  ## Remove generated files
 # -- Test --------------------------------------------------------------
 
 COVERFILE = coverage.out
-COVERAGE = 85.0
+COVERAGE = 82.0
 
 test:  ## Run tests and generate a coverage file
 	go test -covermode=atomic -coverprofile=$(COVERFILE) -race ./...
 	go mod tidy
+
+integration: ## Run integration tests which interact over the network
+	go test -covermode=atomic -tags integration -race ./...
 
 bench: ## Run benchmarks only
 	go test -run=NONE -bench=. ./...
