+++ mod/github_test.go
@@ -29,31 +29,6 @@ func TestGitHubMgrInit(t *testing.T) {
 	assert.Error(t, err)
 }
 
-func TestGitHubMgrGet(t *testing.T) {
-	dir := ".pkgcache"
-	fs := afero.NewMemMapFs()
-	githubmod, err := newGitHubMgr(GitHubOptions{CacheDir: dir, AccessToken: accessTokenForTest(t), Fs: fs})
-	assert.NoError(t, err)
-	testMods := Modules{}
-
-	mod, err := githubmod.Get(RemoteDepsFile, "", &testMods)
-	assert.Nil(t, err)
-	assert.Equal(t, RemoteRepo, mod.Name)
-
-	mod, err = githubmod.Get(RemoteDepsFile, MasterBranch, &testMods)
-	assert.Nil(t, err)
-	assert.Equal(t, RemoteRepo, mod.Name)
-
-	mod, err = githubmod.Get(RemoteDepsFile, "v0.0.1", &testMods)
-	assert.Nil(t, err)
-	assert.Equal(t, RemoteRepo, mod.Name)
-	assert.Equal(t, "v0.0.1", mod.Version)
-
-	mod, err = githubmod.Get("github.com/anz-bank/wrong/path", "", &testMods)
-	assert.Error(t, err)
-	assert.Nil(t, mod)
-}
-
 func TestGitHubMgrFind(t *testing.T) {
 	cacheDir := ".pkgcache"
 	repo := "github.com/foo/bar"
@@ -174,35 +149,6 @@ func TestGetGitHubRepoPath(t *testing.T) {
 			}
 			assert.Equal(t, tt.path, p)
 		})
-	}
-}
-
-func TestGetCacheRef(t *testing.T) {
-	dir := ".pkgcache"
-	githubmod, err := newGitHubMgr(GitHubOptions{CacheDir: dir, AccessToken: accessTokenForTest(t)})
-	assert.NoError(t, err)
-	repoPath := &githubRepoPath{
-		owner: "anz-bank",
-		repo:  "pkg",
-	}
-	ref, err := githubmod.GetCacheRef(repoPath, "v0.0.7")
-	assert.NoError(t, err)
-	assert.Equal(t, "v0.0.7", ref)
-
-	ref, err = githubmod.GetCacheRef(repoPath, MasterBranch)
-	assert.NoError(t, err)
-	assert.Equal(t, "v0.0.0-", ref[:7])
-}
-
-func BenchmarkGetCacheRef(b *testing.B) {
-	dir := ".pkgcache"
-	githubmod, _ := newGitHubMgr(GitHubOptions{CacheDir: dir, AccessToken: accessTokenForTest(b)})
-	repoPath := &githubRepoPath{
-		owner: "anz-bank",
-		repo:  "pkg",
-	}
-	for i := 0; i < b.N; i++ {
-		_, _ = githubmod.GetCacheRef(repoPath, "v0.0.7")
 	}
 }
