+++ mod/gomodules_test.go
@@ -35,28 +35,6 @@ func TestModInitAlreadyExists(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-func TestGoModulesGet(t *testing.T) {
-	gomod := &goModules{}
-	testMods := Modules{}
-
-	mod, err := gomod.Get(RemoteDepsFile, "", &testMods)
-	assert.Nil(t, err)
-	assert.Equal(t, RemoteRepo, mod.Name)
-
-	mod, err = gomod.Get(RemoteDepsFile, MasterBranch, &testMods)
-	assert.Nil(t, err)
-	assert.Equal(t, RemoteRepo, mod.Name)
-
-	mod, err = gomod.Get(RemoteDepsFile, "v0.0.1", &testMods)
-	assert.Nil(t, err)
-	assert.Equal(t, RemoteRepo, mod.Name)
-	assert.Equal(t, "v0.0.1", mod.Version)
-
-	mod, err = gomod.Get("github.com/anz-bank/wrongpath", "", &testMods)
-	assert.Error(t, err)
-	assert.Nil(t, mod)
-}
-
 func TestGoModulesFind(t *testing.T) {
 	gomod := &goModules{}
 	testMods := Modules{}
